--- next.config.js
@@ -1,47 +1,0 @@
-module.exports = {
-  images: {
-    remotePatterns: [
-      {
-        protocol: "https",
-        hostname: "assets.tina.io",
-        port: "",
-      },
-    ],
-  },
-  webpack(config) {
-    config.module.rules.push({
-      test: /\.svg$/i,
-      issuer: /\.[jt]sx?$/,
-      use: ["@svgr/webpack"],
-    });
-
-    return config;
-  },
-  async headers() {
-    // these are also defined in the root layout since github pages doesn't support headers
-    const headers = [
-      {
-        key: "X-Frame-Options",
-        value: "SAMEORIGIN",
-      },
-      {
-        key: "Content-Security-Policy",
-        value: "frame-ancestors 'self'",
-      },
-    ];
-    return [
-      {
-        source: "/(.*)",
-        headers,
-      },
-    ];
-  },
-  async rewrites() {
-    return [
-      {
-        source: "/admin",
-        destination: "/admin/index.html",
-      },
-    ];
-  },
-};
