+++ next.config.ts
@@ -1,0 +1,42 @@
+import type { NextConfig } from 'next'
+ 
+const nextConfig: NextConfig = {
+  images: {
+    remotePatterns: [
+      {
+        protocol: 'https',
+        hostname: 'assets.tina.io',
+        port: '',
+      },
+    ],
+  },
+  async headers() {
+    // these are also defined in the root layout since github pages doesn't support headers
+    const headers = [
+      {
+        key: 'X-Frame-Options',
+        value: 'SAMEORIGIN',
+      },
+      {
+        key: 'Content-Security-Policy',
+        value: "frame-ancestors 'self'",
+      },
+    ];
+    return [
+      {
+        source: '/(.*)',
+        headers,
+      },
+    ];
+  },
+  async rewrites() {
+    return [
+      {
+        source: '/admin',
+        destination: '/admin/index.html',
+      },
+    ];
+  },
+};
+
+export default nextConfig
