+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "tinacms dev -c \"next dev\"",
+    "dev": "tinacms dev -c \"next dev --turbopack\"",
     "build": "tinacms build && next build",
     "build-local": "tinacms build --local --skip-indexing --skip-cloud-checks && next build",
     "start": "tinacms build && next start",
@@ -17,7 +17,7 @@
     "@tinacms/cli": "^1.9.2",
     "@types/js-cookie": "^3.0.6",
     "@types/node": "^22.13.9",
-    "@types/react": "^18.3.18",
+    "@types/react": "19.0.12",
     "@types/styled-components": "^5.1.34",
     "autoprefixer": "^10.4.20",
     "postcss": "^8.5.3",
@@ -27,19 +27,20 @@
   },
   "dependencies": {
     "@headlessui/react": "^2.2.0",
+    "@tailwindcss/postcss": "^4.0.17",
     "@tailwindcss/typography": "^0.5.16",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "date-fns": "^3.6.0",
     "mermaid": "^11.4.1",
-    "next": "^14.2.24",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "next": "15.2.4",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
     "react-icons": "^5.5.0",
     "react-player": "^2.16.0",
     "styled-jsx": "^5.1.6",
     "tailwind-merge": "^2.6.0",
-    "tailwindcss": "^3.4.17",
+    "tailwindcss": "^4.0.17",
     "tinacms": "^2.7.2",
     "usehooks-ts": "^3.1.1"
   }
